feat: use localstorage instead of indexed db for storing last key packages query date
