+++ packages/core/package.json
@@ -11,6 +11,7 @@
     "@wireapp/commons": "4.3.0",
     "@wireapp/cryptobox": "12.8.0",
     "@wireapp/store-engine-dexie": "1.6.10",
+    "@wireapp/store-engine-web-storage": "1.5.11",
     "bazinga64": "5.10.0",
     "hash.js": "1.1.7",
     "http-status-codes": "2.1.4",
