+++ packages/core/src/main/notification/NotificationDatabaseRepository.ts
@@ -22,13 +22,7 @@ import type {Notification} from '@wireapp/api-client/src/notification/';
 import type {CRUDEngine} from '@wireapp/store-engine';
 
 import {CryptographyDatabaseRepository} from '../cryptography/CryptographyDatabaseRepository';
-import {
-  CommonMLS,
-  CompoundGroupIdParams,
-  LastKeyMaterialUpdateParams,
-  LastKeyPackageQueryParams,
-  StorePendingProposalsParams,
-} from './types';
+import {CommonMLS, CompoundGroupIdParams, LastKeyMaterialUpdateParams, StorePendingProposalsParams} from './types';
 
 export enum DatabaseStores {
   EVENTS = 'events',
@@ -168,33 +162,5 @@ export class NotificationDatabaseRepository {
    */
   public async getStoredLastKeyMaterialUpdateDates() {
     return this.storeEngine.readAll<LastKeyMaterialUpdateParams>(STORES.LAST_KEY_MATERIAL_UPDATE_DATES);
-  }
-
-  /**
-   * ## MLS only ##
-   * Store date of last key package query
-   *
-   * @param {lastQueryDate} params.previousUpdateDate - date of the last key packages query
-   */
-  public async storeLastKeyPackageQueryDate(params: LastKeyPackageQueryParams) {
-    await this.storeEngine.updateOrCreate(
-      STORES.LAST_KEY_PACKAGE_QUERY_DATE,
-      DatabaseKeys.LAST_KEY_MATERIAL_QUERY_DATE,
-      params,
-    );
-    return true;
-  }
-
-  /**
-   * ## MLS only ##
-   * Get last key packages query date.
-   *
-   */
-  public async getStoredLastKeyPackagesQueryDate() {
-    const {lastQueryDate} = await this.storeEngine.read<LastKeyPackageQueryParams>(
-      STORES.LAST_KEY_PACKAGE_QUERY_DATE,
-      DatabaseKeys.LAST_KEY_MATERIAL_QUERY_DATE,
-    );
-    return lastQueryDate;
   }
 }
