+++ packages/core/src/main/notification/NotificationService.ts
@@ -41,6 +41,7 @@ import {CommitPendingProposalsParams, HandlePendingProposalsParams, LastKeyMater
 import {TaskScheduler} from '../util/TaskScheduler/TaskScheduler';
 import type {MLSService} from '../mls';
 import {LowPrecisionTaskScheduler} from '../util/LowPrecisionTaskScheduler/LowPrecisionTaskScheduler';
+import {NotificationStorageRepository} from './NotificationStorageRepository';
 
 export type HandledEventPayload = {
   event: Events.BackendEvent;
@@ -72,6 +73,7 @@ export class NotificationService extends EventEmitter {
   private readonly backend: NotificationBackendRepository;
   private readonly cryptographyService: CryptographyService;
   private readonly database: NotificationDatabaseRepository;
+  private readonly storage: NotificationStorageRepository;
   private readonly logger = logdown('@wireapp/core/notification/NotificationService', {
     logger: console,
     markdown: false,
@@ -89,6 +91,7 @@ export class NotificationService extends EventEmitter {
     this.cryptographyService = cryptographyService;
     this.backend = new NotificationBackendRepository(this.apiClient);
     this.database = new NotificationDatabaseRepository(storeEngine);
+    this.storage = new NotificationStorageRepository();
   }
 
   private async getAllNotifications() {
@@ -535,7 +538,7 @@ export class NotificationService extends EventEmitter {
 
     const lastQueryDate = new Date().getTime();
 
-    await this.database.storeLastKeyPackageQueryDate({lastQueryDate});
+    await this.storage.storeLastKeyPackageQueryDate({lastQueryDate});
 
     const minAllowedNumberOfKeyPackages = INITIAL_NUMBER_OF_KEY_PACKAGES / 2;
 
@@ -568,7 +571,7 @@ export class NotificationService extends EventEmitter {
     //if client did not query before, default last date to 0, so it will query the database immediately
     let lastQueryDate = 0;
     try {
-      lastQueryDate = await this.database.getStoredLastKeyPackagesQueryDate();
+      lastQueryDate = await this.storage.getStoredLastKeyPackagesQueryDate();
     } catch (error) {
       this.logger.error('Could not get last key packages query date, using 0 as default value', error);
     }
